Refactor endpoint network execution helpers
